u-net/requirements.txt
# Deep Learning Framework
torch
torchvision

# Image Processing
Pillow

# Scientific Computing
numpy

# Visualization
matplotlib
